% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMark.R
\name{clusterMark}
\alias{clusterMark}
\title{Seurat FindMarkers cluster by cluster and annotate them.}
\usage{
clusterMark(obj, clust, anno, nam)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to use Seurat FindMarkers cluster by cluster and annotate them using an outside reference.
}
\examples{
clusterMark()
}
\keyword{Seurat}
\keyword{cluster,}
\keyword{explore,}
\keyword{scRNA,}
