% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenesNearCoordinateMouse.R
\name{getGenesNearCoordinateMouse}
\alias{getGenesNearCoordinateMouse}
\title{An Annotation Function}
\usage{
getGenesNearCoordinateMouse(annMethResult, regionSize = 50000)
}
\arguments{
\item{regionSize}{How far +/- do you want to search for genes. Defaults to 50,000 bp.}
}
\description{
This function finds all genes surrounding methylated bases +/- a variable region size in mouse.
}
\examples{
getGenesNearCoordinateMouse()
}
\keyword{annotation}
\keyword{mouse,}
