% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHumanOrtho.R
\name{getHumanOrtho}
\alias{getHumanOrtho}
\title{A Ortholog Fetch Function}
\usage{
getHumanOrtho(df = "data frame with refseq mrna id's to convert",
  df_refseq_mrna = "column in dataframe that contains the refseq mrna ids",
  refseq_mrna = "name of column in dataframe that contains the refseq mrna ids as a string")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function gets human ortholog of refseq_mrna ID in rownames.
}
\examples{
getHumanOrtho()
}
\keyword{cats}
