% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGSEA.R
\name{gGSEA}
\alias{gGSEA}
\title{A Gene Set Enrichment Plotting Function}
\usage{
gGSEA(GSEA, db, name, analysis)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to  plot results from WebGestaltR GSEA results
}
\examples{
gGSEA()
}
\keyword{GSEA}
