% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gORA.R
\name{gORA}
\alias{gORA}
\title{A Over-representation Analysis Plotting Function}
\usage{
gORA(ORA, db, name, direction, analysis)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to  plot results from WebGestaltR ORA results
}
\examples{
gORA()
}
\keyword{ORA}
