% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_venn_csv.R
\name{make_venn_csv}
\alias{make_venn_csv}
\title{A Venn Diagram Function}
\usage{
make_venn_csv(VennDiagram_obj, nam, ann, byy)
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a Venn diagram object from VennDiagram package, and makes a CSV file of the overlaps with annotation.
}
\examples{
make_venn_csv()
}
\keyword{venn}
