% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVenns.R
\name{makeVenns}
\alias{makeVenns}
\title{A function to make Venn Diagrams}
\usage{
makeVenns(overlap = "list of elements to overlap",
  nam = "desired file base name",
  cats = "name of each element of your list",
  geneAnno = "gene annotation to add to overlap",
  byy = "what to merge the overlaps and gene annotation by")
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to make 2, 3, or 4 circle Venn diagrams. Outputs images and annotated lists.
}
\examples{
makeVenns()
}
\keyword{venn}
