% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth25bp.R
\name{meth25bp}
\alias{meth25bp}
\title{A Bisulfite Sequencing Function}
\usage{
meth25bp(methylseqObj = "methy seq obj", nam = "name of exp",
  gene.obj = gene.obj, org = "organism name for WebGestaltR",
  db = "GO db's to query")
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a raw methylKit object and does basic QC, then tests for differential methylation in 25bp windows with 15 bp step-size.
}
\examples{
meth25bp()
}
\keyword{cats}
