% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDiffExpDDS.R
\name{myDiffExpDDS}
\alias{myDiffExpDDS}
\title{A Differential Expression Function}
\usage{
myDiffExpDDS(dds, colData1, colData2, formulaEdgR, formulaDE2, ref, name,
  org, anno, db)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you explore basic QC of RNA-seq data and run differential expression testing and GO analysis.
}
\examples{
myDiffExp()
}
\keyword{DESeq2,}
\keyword{QC,}
\keyword{edgeR}
\keyword{explore,}
