% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myExplore.R
\name{myExplore}
\alias{myExplore}
\title{A Data Exploration Function}
\usage{
myExplore(counts, design, colData1, colData2, formulaDE2, ref, name, anno)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you explore basic QC of RNA-seq data without running differential expression testing or GO analysis.
}
\examples{
myExplore()
}
\keyword{DESeq2,}
\keyword{QC,}
\keyword{edgeR}
\keyword{explore,}
