% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGOKmeans_venn.R
\name{myGOKmeans_venn}
\alias{myGOKmeans_venn}
\title{A Venn Diagram GO Analysis Function}
\usage{
myGOKmeans_venn(set_num, counts = NULL, VennDiagram_obj, name, org, db)
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a Venn diagram object from VennDiagram package, and runs WebGestaltR GO annotation.
}
\examples{
myGOKmeans_venn()
}
\keyword{GO,}
\keyword{WebGestalt}
\keyword{venn,}
