% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noAxesNoLegendFeaturePlot.R
\name{noAxesNoLegendFeaturePlot}
\alias{noAxesNoLegendFeaturePlot}
\title{A Seurat Plotting Function}
\usage{
noAxesNoLegendFeaturePlot(p)
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a list of feature plots from Seurat and removes the legends/axes, and plots them in a grid using cowplot.
}
\examples{
noAxesNoLegendFeaturePlot()
}
\keyword{cats}
