% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAnnoTxt.R
\name{processAnnoTxt}
\alias{processAnnoTxt}
\title{A general processing function}
\usage{
processAnnoTxt(rawPath)
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a raw count table generated by Homer's analyze repeats and creates a dataframe of the homer annotation.
}
\examples{
processAnnoTxt()
}
\keyword{cats}
