% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processRawTxt.R
\name{processRawTxt}
\alias{processRawTxt}
\title{A general processing function}
\usage{
processRawTxt(rawPath)
}
\arguments{
\item{no}{defaults}
}
\description{
This function takes in a raw count table generated by Homer's analyze repeats and creates a dataframe with appropriate column names and uses gene names as row names.
}
\examples{
processRawTxt()
}
\keyword{cats}
