% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGSEA.R
\name{rGSEA}
\alias{rGSEA}
\title{A Gene Set Enrichment  Function}
\usage{
rGSEA(gsea.df, org, db, name, analysis)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to preform GSEA analysis
}
\examples{
rGSEA()
}
\keyword{GSEA}
