% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmLaneNum.R
\name{rmLaneNum}
\alias{rmLaneNum}
\title{A Lane Number Removal Function}
\usage{
rmLaneNum(test)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function removes the last 5 characters from any string.
}
\examples{
rmLaneNum()
}
\keyword{five}
