% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmLastElementStrSplit.R
\name{rmLastElementStrSplit}
\alias{rmLastElementStrSplit}
\title{A String Manipulation Function}
\usage{
rmLastElementStrSplit(test)
}
\arguments{
\item{test}{no default}
}
\description{
This function removes the last element of "." or "_" separated string
}
\examples{
rmLastElementStrSplit()
}
\keyword{cats}
