% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rORA.R
\name{rORA}
\alias{rORA}
\title{A Over-representation Analysis Function}
\usage{
rORA(counts, resSig, org, db, name, direction, analysis)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you to preform ORA analysis
}
\examples{
rORA()
}
\keyword{ORA}
