% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Publication.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{A Graphing Function}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function makes ggplot's theme resemble graphpad.
}
\examples{
scale_colour_Publication()
}
\keyword{ggplot}
\keyword{graphing}
\keyword{graphpad}
