% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratQC.R
\name{seuratQC}
\alias{seuratQC}
\title{First steps of scRNA QC and filtering}
\usage{
seuratQC(ids, nam, pmito, nf_max, nf_min)
}
\arguments{
\item{no}{defaults.}
}
\description{
This function allows you explore basic QC of scRNA-seq data.
}
\examples{
seuratQC()
}
\keyword{QC,}
\keyword{Seurat}
\keyword{explore,}
\keyword{scRNA,}
