% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{A String Manipulation Function}
\usage{
simpleCap(x)
}
\arguments{
\item{no}{default}
}
\description{
This function changes case of a word to simple caps. Exp cats -> Cats, or CATS -> Cats.
}
\examples{
simpleCap()
}
\keyword{string}
