% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{A Graphing Function}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function makes ggplot's theme resemble graphpad.
}
\examples{
theme_Publication()
}
\keyword{ggplot}
\keyword{graphing}
\keyword{graphpad}
