% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webG.R
\name{webG}
\alias{webG}
\title{A GO Analysis Function}
\usage{
webG(counts, resSig, org, db, name, gsea.df = NULL)
}
\arguments{
\item{GSEA}{input is default NULL.}
}
\description{
This function takes in gene lists and runs ORA and optionally GSEA analysis on them, against user defined databases.
}
\examples{
webG()
}
\keyword{GO,}
\keyword{GSEA,}
\keyword{ORA,}
\keyword{WebGestalt}
