% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webG_2.R
\name{webG_2}
\alias{webG_2}
\title{A GO Analysis Function 2.0}
\usage{
webG_2(counts, resSig, org, db, name, gsea.df = NULL)
}
\arguments{
\item{GSEA}{input is default NULL.}
}
\description{
This function takes in gene lists and runs ORA and optionally GSEA analysis on them, against user defined databases.
}
\examples{
webG_2()
}
\keyword{GO,}
\keyword{GSEA,}
\keyword{ORA,}
\keyword{WebGestalt}
