% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinClustDE.R
\name{withinClustDE}
\alias{withinClustDE}
\title{A Function to Test for DE WITHIN Seurat Clusters}
\usage{
withinClustDE(seuratObj, clustIdent, groupBy = "orig.ident", identGroup1,
  identGroup2, minPCT = 0.25, geneAnno, nam, org,
  db = c("geneontology_Biological_Process",
  "geneontology_Molecular_Function", "pathway_KEGG"))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to find DE genes between two groups within a cluster, and then run GO analysis.
}
\examples{
withinClustDE()
}
\keyword{DE}
\keyword{scRNA,}
\keyword{seurat,}
\keyword{singlecell,}
